factor out handleFormChange from actions.js
