+++ protocol-designer/src/steplist/actions/index.js
@@ -1,0 +1,9 @@
+// @flow
+import handleFormChange from './handleFormChange'
+
+export * from './actions'
+export * from './types'
+
+export {
+  handleFormChange
+}
