+++ protocol-designer/src/steplist/actions/types.js
@@ -1,0 +1,11 @@
+// @flow
+
+// Update Form input (onChange on inputs)
+export type ChangeFormPayload = {
+  // TODO Ian 2018-05-04 use StepType + FormData type to properly type this payload.
+  // Accessor strings and values depend on StepType
+  stepType?: string,
+  update: {
+    [accessor: string]: string | boolean | Array<string>,
+  }
+}
